adds `param` struct tag (#50)
Given an `Endpointer` struct like this one:

```go
type TaggedStruct struct {
	*Endpoint
	TestParam         string `param:"test_param;a=GET,PUT;r=PUT"`
	TestParamDefault  string `param:"test_param_default"`
	TestParamEmpty    string `param:""`
	TestParamRequired string `param:"test_param_required;a=GET;r=PUT"`
}
```

We apply the following parsing rules:

- `TestParam` has specified `test_param` as the `key` which will map the
incoming input from an HTTP request to the field on the endpoint. It has
also specified `GET` and `PUT` requests to allow the param. Other
requests will strip it out (when that gets implemented). It has
specified `PUT` to require the param. `PUT` requests will throw an error
if `test_param` is not present in the request (when that gets implemented).
- `TestParamDefault` has specified a `key` of `test_param_default`. It
will default to allowing the param on `GET`, `PATCH`, `POST` and `PUT`
requests, and not require it on any of those allowed requests.
- `TestParamEmpty` has not specified any directives, and will use
`TestParamEmpty` as the `key`. It will allow the param on `GET`, `PATCH`,
`POST` and `PUT` requests, which is the default, and not require it on
any of the allowed requests.
- `TestParamRequired` will allow and require `PUT` requests, even though
`PUT` was omitted from the `allowed` list.

fixes #49